requirements.txt
Flask==3.0.3
wheel=0.37.1
python-dotenv==1.0.1
openai==0.28
pandas==2.2.2
openpyxl==3.1.4
uuid==1.30
xlrd==2.0.1